chore: Move error utilities to a new package
Create a new package so callers wishing to use just the error utilities
don't drag in all the other dependencies of `util`.

`LogError` now takes the `Logger` as a parameter, and `client.send` uses
its own logger rather than the Loki global one.
